Handle cleaning stale outputs in one place
