--- platforms/core-execution/execution/src/main/java/org/gradle/internal/execution/steps/RecordOutputsStep.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.internal.execution.steps;
-
-import org.gradle.internal.execution.UnitOfWork;
-import org.gradle.internal.execution.history.OutputFilesRepository;
-
-public class RecordOutputsStep<C extends Context, R extends AfterExecutionResult> implements Step<C, R> {
-    private final OutputFilesRepository outputFilesRepository;
-    private final Step<? super C, ? extends R> delegate;
-
-    public RecordOutputsStep(
-        OutputFilesRepository outputFilesRepository,
-        Step<? super C, ? extends R> delegate
-    ) {
-        this.outputFilesRepository = outputFilesRepository;
-        this.delegate = delegate;
-    }
-
-    @Override
-    public R execute(UnitOfWork work, C context) {
-        R result = delegate.execute(work, context);
-        result.getAfterExecutionState()
-            .ifPresent(afterExecutionState -> outputFilesRepository.recordOutputs(afterExecutionState.getOutputFilesProducedByWork().values()));
-        return result;
-    }
-}
