+++ subprojects/integ-test/src/integTest/groovy/org/gradle/integtests/StaleOutputIntegrationTest.groovy
@@ -18,7 +18,7 @@ package org.gradle.integtests
 
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.BuildOperationsFixture
-import org.gradle.internal.execution.steps.CleanupStaleOutputsStep
+import org.gradle.internal.execution.steps.HandleStaleOutputsStep
 import org.gradle.test.fixtures.file.TestFile
 import org.gradle.util.internal.ToBeImplemented
 import spock.lang.Issue
@@ -294,7 +294,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
 
         then:
         fixture.staleFilesHaveBeenRemoved()
-        operations.hasOperation(CleanupStaleOutputsStep.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
+        operations.hasOperation(HandleStaleOutputsStep.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
     }
 
     def "no build operations are created for stale outputs cleanup if no files are removed"() {
@@ -308,7 +308,7 @@ class StaleOutputIntegrationTest extends AbstractIntegrationSpec {
 
         then:
         executedAndNotSkipped(fixture.taskPath)
-        !operations.first(CleanupStaleOutputsStep.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
+        !operations.first(HandleStaleOutputsStep.CLEAN_STALE_OUTPUTS_DISPLAY_NAME)
     }
 
     def "overlapping outputs between 'build/outputs' and '#overlappingOutputDir' are not cleaned up"() {
